- Made changes for the receiver - Replaced Spectrum Mask with Underlay mask; - inserted Propagation map tab for the receiver.
